backend/requirements.txt
astroid==2.12.13
Click==8.1.3
ecdsa==0.18.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
future==0.18.2
isort==5.11.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
pylint==2.15.8
pycryptodome==3.16.0
python-jose==3.3.0
six==1.16.0
typed-ast==1.5.4
Werkzeug==2.2.2
wrapt==1.14.1
Flask-Cors==3.0.10

